docs(detail): update detail docs and structure
