+++ .storybook/main.js
@@ -15,7 +15,8 @@ module.exports = {
     '../projects/canopy/src/lib/card/card.stories.ts',
     '../projects/canopy/src/lib/carousel/carousel.stories.ts',
     '../projects/canopy/src/lib/data-point/data-point.stories.ts',
-    '../projects/canopy/src/lib/details/details.stories.ts',
+    '../projects/canopy/src/lib/details/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/details/docs/details.stories.ts',
     '../projects/canopy/src/lib/feature-toggle/feature-toggle.stories.ts',
     '../projects/canopy/src/lib/filter-container/filter-container.stories.ts',
     '../projects/canopy/src/lib/forms/checkbox-group/checkbox-group.stories.ts',
