+++ projects/canopy/src/lib/alert/docs/guide.stories.mdx
@@ -38,31 +38,31 @@ There are 5 different variants to choose from:
 
 ![variants](docs/inline-message/variants.png)
 
-### Information
+#### Information
 
 Use to calmly highlight information.
 
 e.g. This product comes with a free umbrella
 
-### Success
+#### Success
 
 Use a success message to tell the user that an action was successful. They require no further action from the user.
 
 e.g. Your umbrella has been dispatched successfully
 
-### Warning
+#### Warning
 
 These messages provide a more elevated level of alert than an information message, and should be used when highlighting interrupted service or something that could affect the user's product holding. They don't necessarily require further action from the user.
 
 e.g. We can't get details of your umbrella right now
 
-### Error
+#### Error
 
 These messages are used to provide feedback that there has been an error that will stop a user from completing their journey. In most cases, this will be when a user enters information on a form which doesn't validate, triggering an error. The content of the message should give concise steps on how to resolve the problem.
 
 e.g. You must confirm the terms & conditions before we can send your umbrella
 
-### Generic
+#### Generic
 
 A basic inline message that should only be used if the others are unsuitable. The icon in this instance is editable and should be tailored to align with the message.
 
@@ -112,7 +112,7 @@ This is based on best practice.
 
 ## Related
 
-- [Details](./?path=/docs/components-details--standard-details)
+- [Details](./?path=/docs/components-details-guide--page)
 - [Message with Brand Icon](./?path=/docs/components-primary-message--standard-primary-message)
 
 ## Help improve this page
